Add UnusuedFileCheck
This check is naive and prone to false positives and false negatives
and hence, as it is, only an optional check.

The check could later be improved by using tree sitter's bash API to
expand e.g. PATCHES, which will probably significantly reduce false
positive rate of this check.

But for now, the check catches most "unused files" for my workflow.

Signed-off-by: Florian Schmaus <flow@gentoo.org>